--- .gitignore
@@ -1,4 +1,0 @@
-plugin
-quick.nvim.tgz
-coc-settings.json
-time.log
