--- .luarc.json
@@ -1,5 +1,0 @@
-{
-    "diagnostics.globals": [
-        "vim"
-    ]
-}
