+++ lua/basics.lua
@@ -1,0 +1,57 @@
+-- Basic settings
+vim.g.mapleader = " "
+vim.g.maplocalleader = " "
+vim.wo.wrap = false
+vim.opt.cursorline = true
+vim.opt.ignorecase = true
+vim.opt.incsearch = true
+vim.opt.mouse = "a"
+vim.opt.number = true
+vim.opt.relativenumber = true
+vim.opt.hlsearch = false
+vim.opt.splitright = true
+vim.opt.splitbelow = true
+vim.opt.termguicolors = true
+vim.opt.signcolumn = "yes"
+vim.opt.hidden = true
+vim.opt.backup = false
+vim.opt.writebackup = false
+vim.opt.swapfile = false
+vim.opt.list = true
+vim.opt.laststatus = 3
+vim.opt.pumheight = 10
+vim.opt.scrolloff = 3
+vim.opt.sidescrolloff = 3
+vim.opt.foldexpr = "nvim_treesitter#foldexpr()"
+vim.opt.foldenable = false
+-- vim.opt.background = "dark"
+
+-- Basic mappings
+vim.keymap.set("n", "<C-S>", ":%s/")
+vim.keymap.set("n", "sp", ":sp<CR>")
+vim.keymap.set("n", "tj", ":tabprev<CR>")
+vim.keymap.set("n", "tk", ":tabnext<CR>")
+vim.keymap.set("n", "tn", ":tabnew<CR>")
+vim.keymap.set("n", "to", ":tabo<CR>")
+vim.keymap.set("n", "vs", ":vs<CR>")
+vim.keymap.set("n", "<leader>j", ":cnext<CR>", { silent = true })
+vim.keymap.set("n", "<leader>k", ":cprevious<CR>", { silent = true })
+vim.keymap.set("n", "<leader>o", ":tabonly<cr>:only<CR>", { silent = true })
+vim.keymap.set("n", "<c-h>", "<c-w>h")
+vim.keymap.set("n", "<c-l>", "<c-w>l")
+vim.keymap.set("n", "<c-j>", "<c-w>j")
+vim.keymap.set("n", "<c-k>", "<c-w>k")
+
+-- Auto resize splits when the terminal's window is resized
+vim.api.nvim_create_autocmd({ "VimResized" }, {
+	group = vim.api.nvim_create_augroup("EqualizeSplits", {}),
+	callback = function()
+		local current_tab = vim.api.nvim_get_current_tabpage()
+		vim.cmd("tabdo wincmd =")
+		vim.api.nvim_set_current_tabpage(current_tab)
+	end,
+	desc = "Resize splits with terminal window",
+})
+
+-- Global LSP mappings
+vim.keymap.set("n", "<space>e", vim.diagnostic.open_float)
