+++ lua/plugins/autopairs.lua
@@ -1,0 +1,7 @@
+return {
+	"windwp/nvim-autopairs",
+	event = "InsertEnter",
+	opts = {
+		disable_filetype = { "TelescopePrompt", "vim" },
+	},
+}
