+++ lua/plugins/autotag.lua
@@ -1,0 +1,4 @@
+return {
+	"windwp/nvim-ts-autotag",
+	config = true,
+}
