+++ lua/plugins/init.lua
@@ -1,0 +1,5 @@
+return {
+	"nvim-lua/plenary.nvim",
+	"tpope/vim-sleuth",
+	"tpope/vim-rhubarb",
+}
