+++ lua/plugins/surround.lua
@@ -1,0 +1,5 @@
+return {
+	"kylechui/nvim-surround",
+	event = "VeryLazy",
+	opts = {},
+}
