+++ lua/plugins/treesitter.lua
@@ -1,0 +1,12 @@
+return {
+	"nvim-treesitter/nvim-treesitter",
+	build = ":TSUpdate",
+	config = function()
+		local configs = require("nvim-treesitter.configs")
+
+		configs.setup({
+			highlight = { enable = true },
+			indent = { enable = true },
+		})
+	end,
+}
