+++ lua/plugins/ts-context-commentstring.lua
@@ -1,0 +1,6 @@
+return {
+	"JoosepAlviste/nvim-ts-context-commentstring",
+	opts = {
+		enable_autocmd = false,
+	},
+}
